Made some changes to setting totalPrice on the order in terms of trying to fix totalcost, though this was not related to the totalcost bug.

+++ src/java/uts/isd/model/OrderLine.java
@@ -114,7 +114,7 @@ public class OrderLine {
     }
     
     public double getPrice() {
-        return this.unitPrice * this.quantity;
+        return ((double)this.unitPrice * (double)this.quantity);
     }
    
     public String getPriceFormatted(String format) {
